fix os dependent bash scripts

+++ install.sh
@@ -1,11 +1,22 @@
-curl https://nixos.org/nix/install | sh
-. ~/.nix-profile/etc/profile.d/nix.sh
+if test -f /etc/NIXOS; then
+    echo "" 
+else
+    curl https://nixos.org/nix/install | sh 
+
+    . ~/.nix-profile/etc/profile.d/nix.sh
+fi
 
 ./link-configs.sh
 
-defaults write com.apple.Dock autohide-delay -float 5 && killall Dock
-defaults write -g ApplePressAndHoldEnabled -bool false
-
 nix-env -i all
 
-chsh -s "$(which zsh)"
+if [[ "$OSTYPE" == "darwin"* ]]; then
+    defaults write com.apple.Dock autohide-delay -float 5 && killall Dock
+    defaults write -g ApplePressAndHoldEnabled -bool false
+fi
+
+if test -f /etc/NIXOS; then
+    echo ""
+else
+    chsh -s "$(which zsh)"
+fi
